Move symbolic distributions definitions to their own file
