+++ src/pharmpy/model/__init__.py
@@ -1,16 +1,10 @@
 from .data import DatasetError, DatasetWarning
 from .datainfo import ColumnInfo, DataInfo
+from .distributions.symbolic import Distribution, JointNormalDistribution, NormalDistribution
 from .estimation import EstimationStep, EstimationSteps
 from .model import Model, ModelError, ModelfitResultsError, ModelSyntaxError
 from .parameters import Parameter, Parameters
-from .random_variables import (
-    Distribution,
-    JointNormalDistribution,
-    NormalDistribution,
-    RandomVariables,
-    VariabilityHierarchy,
-    VariabilityLevel,
-)
+from .random_variables import RandomVariables, VariabilityHierarchy, VariabilityLevel
 from .results import Results
 from .statements import (
     Assignment,
